+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Update
 
+- Improve View Header, use less space
 - Improve mobile support when editing text field
 - Only display the segment selection '✓' symbol on larger screen